backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
httpx>=0.26.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
google-auth>=2.28.0
requests>=2.31.0
PyJWT>=2.8.0
